+++ src/nbt/ShortTag.cpp
@@ -19,7 +19,7 @@ Tag::Type ShortTag::getType() const { return Type::Short; }
 
 std::unique_ptr<Tag> ShortTag::copy() const { return std::make_unique<ShortTag>(mStorage); }
 
-std::size_t ShortTag::hash() const { return mStorage; }
+std::size_t ShortTag::hash() const { return static_cast<size_t>(mStorage); }
 
 void ShortTag::write(BytesDataOutput& stream) const { stream.writeShort(mStorage); }
